requirements.txt
certifi==2021.5.30
chardet==4.0.0
click==8.0.1
DBUtils==1.3
faiss-gpu==1.7.1
Flask==2.0.1
idna==2.10
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
marshmallow==3.12.1
numpy==1.20.3
PyMySQL==1.0.2
requests==2.25.1
toml==0.10.2
typing-extensions==3.10.0.0
urllib3==1.26.5
waitress==2.0.0
Werkzeug==2.0.1
